refactor: merge Paramore.Darker.Testing.Ports into Tests/Exported (#321)
Tidy First (structural-only): fold the 5 source files from
`test/Paramore.Darker.Testing.Ports/` into
`test/Paramore.Darker.Tests/Exported/` with namespace
`Paramore.Darker.Core.Tests.Exported` (the post-Step-8 target). Flip the
14 public class declarations in `TestDoubles/*.cs` to `internal`, plus
the 4 nested public test handlers in `Decorators/FallbackPolicyTests.cs`
that were polluting `AddHandlersFromAssemblies` scans. Retarget the AOT
csproj's `<ProjectReference>` to `Paramore.Darker.Tests.csproj` with
`<PrivateAssets>all</PrivateAssets>`. Drop Tests's reference to
`Paramore.Darker.Testing.Ports`. Update 9 `using` statements.

AOT test count is AOT-only (6 tests on net8.0 and net9.0); full filter
suite passes (77 + 77 + 6 + 6 = 166 tests).
