+++ test/Paramore.Darker.Tests.AOT/Paramore.Darker.Tests.AOT.csproj
@@ -10,7 +10,9 @@
     <ProjectReference Include="..\..\src\Paramore.Darker.QueryLogging\Paramore.Darker.QueryLogging.csproj" />
     <ProjectReference Include="..\..\src\Paramore.Darker.Testing\Paramore.Darker.Testing.csproj" />
     <ProjectReference Include="..\..\src\Paramore.Darker\Paramore.Darker.csproj" />
-    <ProjectReference Include="..\Paramore.Darker.Testing.Ports\Paramore.Darker.Testing.Ports.csproj" />
+    <ProjectReference Include="..\Paramore.Darker.Tests\Paramore.Darker.Tests.csproj">
+      <PrivateAssets>all</PrivateAssets>
+    </ProjectReference>
     <ProjectReference Include="..\Paramore.Test.Helpers\Paramore.Test.Helpers.csproj" />
   </ItemGroup>
   <ItemGroup>
