+++ test/Paramore.Darker.Tests.AOT/QueryProcessor/AOTQueryProcessorTests.cs
@@ -1,6 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Paramore.Darker.Testing.Ports;
+using Paramore.Darker.Core.Tests.Exported;
 using Shouldly;
 using System;
 using System.Threading.Tasks;
