+++ test/Paramore.Darker.Tests/FakeQueryProcessorTests.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Paramore.Darker.Testing;
-using Paramore.Darker.Testing.Ports;
+using Paramore.Darker.Core.Tests.Exported;
 using Shouldly;
 using Xunit;
