+++ test/Paramore.Darker.Tests/Integrations/When_AddHandlersFromAssemblies_scans_assembly_should_register_both_sync_and_async_handlers.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Paramore.Darker.Extensions.DependencyInjection;
-using Paramore.Darker.Testing.Ports;
+using Paramore.Darker.Core.Tests.Exported;
 using Shouldly;
 using Xunit;
