+++ test/Paramore.Darker.Tests/Paramore.Darker.Tests.csproj
@@ -9,7 +9,6 @@
     <ProjectReference Include="..\..\src\Paramore.Darker.QueryLogging\Paramore.Darker.QueryLogging.csproj" />
     <ProjectReference Include="..\..\src\Paramore.Darker.Testing\Paramore.Darker.Testing.csproj" />
     <ProjectReference Include="..\..\src\Paramore.Darker\Paramore.Darker.csproj" />
-    <ProjectReference Include="..\Paramore.Darker.Testing.Ports\Paramore.Darker.Testing.Ports.csproj" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
