+++ test/Paramore.Darker.Tests/QueryHandlerRegistryTests.cs
@@ -1,6 +1,6 @@
 using System;
 using Paramore.Darker.Exceptions;
-using Paramore.Darker.Testing.Ports;
+using Paramore.Darker.Core.Tests.Exported;
 using Shouldly;
 using Xunit;
