+++ test/Paramore.Darker.Tests/QueryProcessorTests.cs
@@ -1,6 +1,6 @@
 using System;
 using Moq;
-using Paramore.Darker.Testing.Ports;
+using Paramore.Darker.Core.Tests.Exported;
 using Shouldly;
 using Xunit;
