+++ test/Paramore.Darker.Tests/TestDoubles/LoggingQueryHandler.cs
@@ -2,7 +2,7 @@ using Paramore.Darker.Logging.Attributes;
 
 namespace Paramore.Darker.Tests.TestDoubles
 {
-    public class LoggingQueryHandler : QueryHandler<SyncTestQuery, SyncTestQuery.Result>
+    internal class LoggingQueryHandler : QueryHandler<SyncTestQuery, SyncTestQuery.Result>
     {
         [QueryLogging(1)]
         public override SyncTestQuery.Result Execute(SyncTestQuery query)
