+++ test/Paramore.Darker.Tests/When_QueryProcessorBuilder_builds_processor_should_configure_both_sync_and_async.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Threading.Tasks;
 using Paramore.Darker.Builder;
-using Paramore.Darker.Testing.Ports;
+using Paramore.Darker.Core.Tests.Exported;
 using Shouldly;
 using Xunit;
